Start of map development
